Hack a fix for the year getting displayed as the ID
My data is just formatted differently. A nicer fix would be to
format my data as this thing expects. Or nicer still would be a
better separation of concerns. But this will do.

I guess the first year is for the graph and the second year
(previously date) is for the table -- but that is just a guess.

Still don't know why everything gets listed twice in the table.