In embeddings view, control topic via URL param
This allows me to locate topics in this view, albeit only by
typing the topic ID into the URL.

But creates potential to add links to embeddings views that at
least start with the correct topic.

And gives me an example for how to do this kind of thing.

Note that the colouration is fairly subtle, so easy to miss for
small circles.

Also need to check whether it works on first visit, or requires a
refresh.